refactor openBrowser
